+++ src/exceptions/EmptyTableException.php
